Se corrigio el delivery y se agregó el endpoint de cancelación y actualización de estos

+++ SmartOrderService/SmartOrderService/Models/Requests/InventoryRequest.cs
@@ -46,7 +46,7 @@ namespace SmartOrderService.Models.Requests
     {
         public int productId { get; set; }
         public int quantity { get; set; }
-        public double price { get; set; }
+        public Single price { get; set; }
     }
     #endregion
 }
